+++ src/snippets/snippetSyntaxTs.ts
@@ -1,0 +1,10 @@
+/**
+ * SnippetSyntax class for typescript.
+ * Important to use separate file as it's simplest way to distinguish between js/ts
+ * without hooking into cucumber machinery.
+ */
+import SnippetSyntax from './snippetSyntax';
+
+export default class extends SnippetSyntax {
+  isTypescript = true;
+}