+++ test/undefined-step/steps.ts
@@ -1,8 +1,0 @@
-import { expect } from '@playwright/test';
-import { createBdd } from '../../dist';
-
-const { Given } = createBdd();
-
-Given('some state', async ({ page }) => {
-  expect(page).toBeDefined();
-});