+++ test/undefined-steps-cucumber-snippet/playwright.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from '@playwright/test';
+import { defineBddConfig } from '../../dist';
+import { SnippetInterface } from '@cucumber/cucumber/lib/formatter/step_definition_snippet_builder/snippet_syntax';
+
+const testDir = defineBddConfig({
+  paths: ['*.feature'],
+  require: ['steps.ts'],
+  requireModule: ['ts-node/register'],
+  formatOptions: {
+    snippetInterface: SnippetInterface.AsyncAwait,
+  },
+});
+
+export default defineConfig({
+  testDir,
+  forbidOnly: Boolean(process.env.FORBID_ONLY),
+});