+++ test/undefined-steps-cucumber-snippet/steps.ts
@@ -1,0 +1,5 @@
+import { Given } from '@cucumber/cucumber';
+
+Given('some state', async function () {
+  // ...
+});