+++ test/undefined-steps-playwright-snippet/steps.ts
@@ -1,0 +1,7 @@
+import { createBdd } from '../../dist';
+
+const { Given } = createBdd();
+
+Given('some state', async ({}) => {
+  // ...
+});